feat: Support encoding/emitting nested structures with binary fields
I've pulled in jsoniter to support passing a context into a json serialization, so we can stop walking and rewriting the entire passed structure to replace with placeholders.  Instead, for the registered types, we replace them with placeholders inline as they're serialized, writing their attachments to a stream-context-passed attachments list, so it's multiple-simultaneous-encoding-safe.